+++ ServerBrowser/Games/GameExtension.cs
@@ -299,6 +299,8 @@ namespace ServerBrowser
 
     public void Add(Game game, GameExtension extension)
     {
+      if (extension == null)
+        throw new ArgumentNullException(nameof(extension));
       extensions[game] = extension;
       extension.Steamworks = this.Steamworks;
     }
@@ -311,6 +313,8 @@ namespace ServerBrowser
         extension = new GameExtension();
         this.Add(game, extension);
       }
+      if (extension == null)
+        throw new InvalidOperationException("foo");
       return extension;
     }